pyproject.toml
[project]
name = "ideation-agentic-system"
version = "0.1.0"
description = "Discord AI assistant powered by Gemini 2.5 Flash and LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0",
    "discord-py>=2.6.4",
    "google-api-python-client>=2.120.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.27.0",
    "langchain>=1.2.10",
    "langchain-google-genai>=2.0.0",
    "langgraph>=1.0.9",
    "pypdf>=4.0.0",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.5.0",
    "trafilatura>=1.12.0",
]

[project.scripts]
ideation-system = "discord_bot.main:main"
